View точки маршрута умеет принимать данные - В класс EventView добавлен конструктор, в который передаются и сохраняются данные - Передаём данные в шаблон event.tpl.js, создав параметр для функции createEventTemplate - Вместо статичных данных в шаблон выведены моковые данные: даты события, тип события, город, базовая стоимость, отметку в избранном маршрут или нет - Используем библиотеку day.js для вывода даты точки маршрута в человеческом формате. Не учитываем момент, что дата может отсутствовать. - Сделаем на будущее функцию (isPastDate), проверяющую просрочена ли передаваемая дата.
