+++ package.json
@@ -28,5 +28,8 @@
   },
   "engines": {
     "node": "16"
+  },
+  "dependencies": {
+    "dayjs": "1.11.1"
   }
 }
