+++ src/utils.js
@@ -1,3 +1,5 @@
+import dayjs from 'dayjs';
+
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 const getRandomInteger = (a = 0, b = 1) => {
@@ -23,4 +25,24 @@ const getRandomDescription = () => loremText.substring(0, (getRandomInteger(1, 6
 //Функция получения строки случайной длины для полей title
 const getRandomTitle = () => loremText.substring(0, (getRandomInteger(1, 6) * 25));
 
-export {getRandomInteger, commonType, getCity, getRandomDescription, getRandomTitle};
+
+const MIN_BASE_PRICE = 300;
+const MAX_BASE_PRICE = 1100;
+const MIN_MONTH_VALUE = 4;
+const MIDDLE_MONTH_VALUE = 5;
+const MAX_MONTH_VALUE = 6;
+
+const getBasePrice = () => getRandomInteger(MIN_BASE_PRICE, MAX_BASE_PRICE);
+const getMonth = () => [getRandomInteger(MIN_MONTH_VALUE, MIDDLE_MONTH_VALUE), getRandomInteger(MIDDLE_MONTH_VALUE, MAX_MONTH_VALUE)];
+
+//Функция преобразующая дату из формата '2022-04-29T15:59:39.600Z' в формат 'Apr 29'
+const humanizeDateMonthDay = (anyDate) => dayjs(anyDate).format('MMM DD');
+
+//Функция которая помогает определить, передаваемая дата (день) раньше текущего или нет
+//Пример, сегодня следующая дата: '2022-05-02T15:59:39.600Z'
+//Передавая '2022-06-29T15:59:39.600Z' получим 'false'
+//Передавая '2022-03-29T15:59:39.600Z' получим 'true'
+const isPastDate = (anyDate) => anyDate && dayjs().isAfter(anyDate, 'D');
+
+export {getRandomInteger, commonType, getCity, getRandomDescription, getRandomTitle, humanizeDateMonthDay, getBasePrice,
+  getMonth, isPastDate};
